Moved mkldnn conv availability checks to utils and use it across passes (#1984)
* Moved mkldnn conv availability checks to utils and use it across passes

* Style fix
